move builder out of component_string
