+++ src/component_string.rs
@@ -1,3 +1,4 @@
+use crate::template_builder::BuilderImpl;
 use crate::components::Component;
 use crate::compose_steps::{
     compose_steps, push_attr_component, push_attr_value_component, push_text_component,
@@ -5,7 +6,7 @@ use crate::compose_steps::{
 use crate::routes::StepKind;
 use crate::rulesets::RulesetImpl;
 use crate::tag_info::TagInfo;
-use crate::template_steps::{compose, Results as TemplateSteps};
+use crate::template_steps::Results as TemplateSteps;
 
 struct TemplateBit {
     pub inj_index: usize,
@@ -16,25 +17,6 @@ enum StackBit<'a> {
     Tmpl(&'a Component, TemplateSteps, TemplateBit),
     Cmpnt(&'a Component),
     None,
-}
-
-pub trait BuilderImpl {
-    fn build(&mut self, rules: &dyn RulesetImpl, template_str: &str) -> TemplateSteps;
-}
-
-pub struct Builder {}
-
-impl Builder {
-    pub fn new() -> Builder {
-        Builder {}
-    }
-}
-
-impl BuilderImpl for Builder {
-    fn build(&mut self, rules: &dyn RulesetImpl, template_str: &str) -> TemplateSteps {
-        // chance to cache templates here
-        compose(rules, template_str)
-    }
 }
 
 pub fn compose_string(
